plscope_identifiers: make SQL stmt appear wt the EXECUTE usage
Reason: required for pull-request.

I'd argue that EXECUTE is not the best choice of a pseudo-usage
name for SQL statements--we're not executing anything, this is
static code analysis, and all about usages of identifiers/SQL
statements... But it's been EXECUTE so far, and that change is
not ready to be merged yet.