removed filter on origin_con_id
reason: In an multitenant environment the root container contains
some objects that are compiled with PL/Scope. These are the
PL/SQL packages STANDARD and DBMS_STANDARD and the
public synonym DBMS_STANDARD. This means beside the
sys_context('USERENV', 'CON_ID') the CON_ID 1 is also relevant.
Excluding the root would might lead to additional broken hierarchies
which are unnecessarily handled later on. In the end the result
might be incomplete, especially if one would like to analyze the
usage of objects in the root.