"Compile with PL/Scope...": check plsql_ccflags before recomp.
Recompilation in the "Compile with PL/Scope..." action does not
preserve prior compilation settings (they need to be changed in
order to change plscope_settings, to begin with). As a consequence,
if the whole schema is being recompiled in the action, compilation
settings will be changed and made uniform accross the entire schema:
pre-existing per-module settings, if any, will be lost. Obviously
this may not be a good thing for modules which have a value set
for plsql_ccflags, as such flags are set on a per-module basis in
general. This is a kind of data loss we'd rather avoid, and for
that reason, at this stage it's safer not to recompile PL/SQL code
if use of plsql_ccflags is detected in the target schema.