"Compile with PL/Scope...": warn if per-module pl/sql settings
This adds a warning if recompiling the code in the target schema
would make PL/SQL compilation settings uniform accross the whole
schema whereas differences between modules exist before recompiling.
In that case, the default value for the Compile Code? prompt is
changed from Yes to No.

Of course, it would be far better to rewrite the whole compilation
procedure, ensuring that we retain prior settings for each module.
But we'll save that for later.