"Compile with PL/Scope...": added log to dbms_output (optional)
This provides with an optional readout containing a summary of
processing steps, plus (optionally) the list of failed statements,
or the list of all statements ran by the action. The readout is
emitted through dbms_output, hence it is subject to the limitation
of SQL Developer as regards using dbms_output in actions, namely:
SQL Developer does not read from the output buffer after the action
completes, triggering that read is left to the user.

(Running exec null; in a SQL worksheet, if serveroutput is on, or if
the Dbms Output pane is enabled for the user's session, is enough to
have SQL Developer read from the buffer and print its contents.)

Also made ORA-01031 exceptions to be expected/normal under most
circumstances; previously they were expected only when recompiling
public synonyms. Yet it turns out, rather surprisingly, that they
may also happen when recompiling other objects, for reasons that
not necessarily obvious.